+++ lib/TWCManager/Control/HTTPControl.py
@@ -66,6 +66,7 @@ def CreateHTTPHandlerClass(master):
     class HTTPControlHandler(BaseHTTPRequestHandler):
         ampsList = []
         fields = {}
+        host = None
         hoursDurationList = []
         master = None
         path = ""
@@ -135,6 +136,7 @@ def CreateHTTPHandlerClass(master):
             self.templateEnv.globals.update(
                 getMFADevices=master.getModuleByName("TeslaAPI").getMFADevices
             )
+            self.templateEnv.globals.update(host=self.host)
             self.templateEnv.globals.update(hoursDurationList=self.hoursDurationList)
             self.templateEnv.globals.update(navbarItem=self.navbar_item)
             self.templateEnv.globals.update(optionList=self.optionList)
@@ -663,6 +665,16 @@ def CreateHTTPHandlerClass(master):
         def do_GET(self):
             self.url = urllib.parse.urlparse(self.path)
 
+            # Determine host header entry
+            try:
+                self.host = self.headers.get("Host", "")
+
+                # Remove port number from domain
+                if ':' in self.host:
+                    self.host = self.host.split(':',1)[0]
+            except IndexError:
+                self.host = None
+
             # serve local static content files (from './lib/TWCManager/Control/static/' dir)
             if self.url.path.startswith("/static/"):
                 content_type = mimetypes.guess_type(self.url.path)[0]
@@ -890,8 +902,9 @@ def CreateHTTPHandlerClass(master):
 
             if self.url.path == "/teslaAccount/submitCaptcha":
                 captchaCode = self.getFieldValue("captchaCode")
+                interface = self.getFieldValue("interface")
 
-                resp = master.getModuleByName("TeslaAPI").submitCaptchaCode(captchaCode)
+                resp = master.getModuleByName("TeslaAPI").submitCaptchaCode(captchaCode, interface)
 
                 self.send_response(302)
                 self.send_header("Location", "/teslaAccount/" + str(resp))
@@ -1176,11 +1189,17 @@ def CreateHTTPHandlerClass(master):
                     carapi = master.getModuleByName("TeslaAPI")
                     if key == "carApiBearerToken":
                         carapi.setCarApiBearerToken(self.getFieldValue(key))
+                        # We don't know the token expiry time as it was entered manually,
+                        # but we'll assume it was freshly created which means 45 day expiry
+                        carapi.setCarApiTokenExpireTime(time.time() + 45 * 24 * 60 * 60)
                     elif key == "carApiRefreshToken":
                         carapi.setCarApiRefreshToken(self.getFieldValue(key))
+                        carapi.setCarApiTokenExpireTime(time.time() + 45 * 24 * 60 * 60)
 
-                # Write setting to dictionary
-                master.settings[key] = self.getFieldValue(key)
+                else:
+
+                    # Write setting to dictionary
+                    master.settings[key] = self.getFieldValue(key)
 
             # If Non-Scheduled power action is either Do not Charge or
             # Track Green Energy, set Non-Scheduled power rate to 0