+++ CHANGELOG.md
@@ -10,8 +10,10 @@ This document logs the changes per release of TWCManager.
       * (@MikeBishop) - Fix debounce mechanism to avoid continuous charging when minimum generation is not met in some circumstances
       * (@ngardiner) - Align web interface TWCID so that it is lowercase, avoiding issues with JavaScript and Jinja2 misalignment
       * (@ngardiner) - Change cryptography module dependency to avoid versions requiring rust compiler
+      * (@ngardiner) - Add logic to set token expiry time when manually adding tokens
  * (@MikeBishop) - Change to Powerwall Auth handling to expire auth tokens earlier due to changes in Powerwall auth handling
  * (@ngardiner) - Add version check to Web UI
+ * (@ngardiner) - Add handling of recaptcha challenge (with DNS workaround), and provide other options if this is not possible
 
 ## v1.2.3 - 2021-08-10